+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 1.1.43
 
 -   [macros expand on hover when in keys for a mapping, but not as values #278](https://github.com/jlchmura/lpc-language-server/issues/278)
+-   [Macro in union type does not show quickinfo on hover #290](https://github.com/jlchmura/lpc-language-server/issues/290)
 
 ## 1.1.42